+++ docs/html/group__init.js
@@ -1,5 +1,6 @@
 var group__init =
 [
+    [ "SdsRawMessageCallback", "group__init.html#gadf7c1e7d63a48b93e9ddc2e8a3d87a2e", null ],
     [ "sds_get_node_id", "group__init.html#gab2d009b20c5946bae1e107a1efefee0f", null ],
     [ "sds_get_stats", "group__init.html#gab9070cd3bc84e1f14e681e7d626edbb2", null ],
     [ "sds_init", "group__init.html#ga99eea78bf2771e604ab5ce296cd8a815", null ],
@@ -7,5 +8,7 @@ var group__init =
     [ "sds_is_ready", "group__init.html#ga3fb2f5c3f803258542228416cad2cea6", null ],
     [ "sds_loop", "group__init.html#ga275b11badbc4756a9b93502af7e30778", null ],
     [ "sds_publish_raw", "group__init.html#ga1fcaa6529e839b694390d2cfceb77b56", null ],
-    [ "sds_shutdown", "group__init.html#gae809cf426b23805d83699f84d67882ca", null ]
+    [ "sds_shutdown", "group__init.html#gae809cf426b23805d83699f84d67882ca", null ],
+    [ "sds_subscribe_raw", "group__init.html#gac05ed532be503ce8e3be0b073ea08ec5", null ],
+    [ "sds_unsubscribe_raw", "group__init.html#gaafa26c698888d3125187984d75b3bd31", null ]
 ];