+++ docs/html/search/all_13.js
@@ -4,6 +4,6 @@ var searchData=
   ['license_1',['License',['../index.html#autotoc_md34',1,'']]],
   ['linux_20macos_2',['C Library (Linux/macOS)',['../index.html#autotoc_md22',1,'']]],
   ['liveness_5finterval_5fms_3',['liveness_interval_ms',['../struct_sds_table_meta.html#a55a9040b9fad25fdbe3d30f7f5033ecb',1,'SdsTableMeta']]],
-  ['logging_20diagnostics_4',['9. Raw MQTT Publish (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]],
+  ['logging_20diagnostics_4',['9. Raw MQTT Publish/Subscribe (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]],
   ['loop_5',['loop',['../class_s_d_s_client.html#aaa368a52bd7533c87f5b72857e0859d9',1,'SDSClient']]]
 ];