+++ docs/html/search/all_14.js
@@ -4,7 +4,7 @@ var searchData=
   ['messages_5freceived_1',['messages_received',['../struct_sds_stats.html#a52decfc1a4d53f109081ece0b3a33893',1,'SdsStats']]],
   ['messages_5fsent_2',['messages_sent',['../struct_sds_stats.html#a1d631a8649de52256a54a9c04e31010c',1,'SdsStats']]],
   ['mqtt_20broker_3',['5. Start an MQTT Broker',['../index.html#autotoc_md12',1,'']]],
-  ['mqtt_20publish_20logging_20diagnostics_4',['9. Raw MQTT Publish (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]],
+  ['mqtt_20publish_20subscribe_20logging_20diagnostics_4',['9. Raw MQTT Publish/Subscribe (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]],
   ['mqtt_5fbroker_5',['mqtt_broker',['../struct_sds_config.html#af0aad86b3113014cc79fc99044b96721',1,'SdsConfig']]],
   ['mqtt_5fpassword_6',['mqtt_password',['../struct_sds_config.html#a29862638ca158cc4a5681219ee668da6',1,'SdsConfig']]],
   ['mqtt_5fport_7',['mqtt_port',['../struct_sds_config.html#a2f9b1c8ff312e5e48d36cdc3f748d295',1,'SdsConfig']]],