+++ docs/html/search/pages_13.js
@@ -3,5 +3,5 @@ var searchData=
   ['library_20linux_20macos_0',['C Library (Linux/macOS)',['../index.html#autotoc_md22',1,'']]],
   ['license_1',['License',['../index.html#autotoc_md34',1,'']]],
   ['linux_20macos_2',['C Library (Linux/macOS)',['../index.html#autotoc_md22',1,'']]],
-  ['logging_20diagnostics_3',['9. Raw MQTT Publish (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]]
+  ['logging_20diagnostics_3',['9. Raw MQTT Publish/Subscribe (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]]
 ];