+++ docs/html/search/pages_14.js
@@ -2,5 +2,5 @@ var searchData=
 [
   ['macos_0',['C Library (Linux/macOS)',['../index.html#autotoc_md22',1,'']]],
   ['mqtt_20broker_1',['5. Start an MQTT Broker',['../index.html#autotoc_md12',1,'']]],
-  ['mqtt_20publish_20logging_20diagnostics_2',['9. Raw MQTT Publish (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]]
+  ['mqtt_20publish_20subscribe_20logging_20diagnostics_2',['9. Raw MQTT Publish/Subscribe (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]]
 ];