+++ docs/html/search/pages_16.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['publish_20logging_20diagnostics_0',['9. Raw MQTT Publish (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]],
+  ['publish_20subscribe_20logging_20diagnostics_0',['9. Raw MQTT Publish/Subscribe (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]],
   ['python_1',['Python',['../index.html#autotoc_md13',1,'6. Run Your First Device (Python)'],['../index.html#autotoc_md14',1,'7. Run an Owner (Python)']]]
 ];