+++ docs/html/search/pages_8.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['9_20raw_20mqtt_20publish_20logging_20diagnostics_0',['9. Raw MQTT Publish (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]]
+  ['9_20raw_20mqtt_20publish_20subscribe_20logging_20diagnostics_0',['9. Raw MQTT Publish/Subscribe (Logging, Diagnostics)',['../index.html#autotoc_md16',1,'']]]
 ];