+++ docs/html/search/typedefs_0.js
@@ -5,8 +5,9 @@ var searchData=
   ['sdsdeviceevictedcallback_2',['SdsDeviceEvictedCallback',['../group__callbacks.html#ga1620244dfc15a6423fe29b7d9c3edc3e',1,'sds.h']]],
   ['sdserrorcallback_3',['SdsErrorCallback',['../group__callbacks.html#ga0c01cb5dd7ba9f1283e33dfac541fe08',1,'sds.h']]],
   ['sdsnodeiterator_4',['SdsNodeIterator',['../group__callbacks.html#ga663545907da5cc9b3981185ffe824c52',1,'sds.h']]],
-  ['sdsserializefunc_5',['SdsSerializeFunc',['../group__serialization.html#ga7c07d1ba6575c0a6a05017bf410d3d6c',1,'sds.h']]],
-  ['sdsstatecallback_6',['SdsStateCallback',['../group__callbacks.html#ga6733edea0d33e8c1718333a493cdd254',1,'sds.h']]],
-  ['sdsstatuscallback_7',['SdsStatusCallback',['../group__callbacks.html#gab350835fef85e82bb47c8bc41f931e09',1,'sds.h']]],
-  ['sdsversionmismatchcallback_8',['SdsVersionMismatchCallback',['../group__callbacks.html#ga5ed0d3421d5a6a90667376da7b1dd7c0',1,'sds.h']]]
+  ['sdsrawmessagecallback_5',['SdsRawMessageCallback',['../group__init.html#gadf7c1e7d63a48b93e9ddc2e8a3d87a2e',1,'sds.h']]],
+  ['sdsserializefunc_6',['SdsSerializeFunc',['../group__serialization.html#ga7c07d1ba6575c0a6a05017bf410d3d6c',1,'sds.h']]],
+  ['sdsstatecallback_7',['SdsStateCallback',['../group__callbacks.html#ga6733edea0d33e8c1718333a493cdd254',1,'sds.h']]],
+  ['sdsstatuscallback_8',['SdsStatusCallback',['../group__callbacks.html#gab350835fef85e82bb47c8bc41f931e09',1,'sds.h']]],
+  ['sdsversionmismatchcallback_9',['SdsVersionMismatchCallback',['../group__callbacks.html#ga5ed0d3421d5a6a90667376da7b1dd7c0',1,'sds.h']]]
 ];